Main UI Progress
Set up simple buttons for main page UI, definitely needs to be worked on. Set up the functionality to go back and forth between pages with clicking on the New Expense button and displaying a new expense page, and going back to the main page. Simple stuff, but this is different than creating a desktop application. Some things to think about will be handling the Android back button. But no need to get ahead of myself now.