New Income Page changes
Finished adding all buttons and logic to the new income page. BUT, now the new income page has the functionality to be a single record page, and the need for separate income and expense input pages is no longer there. These files will be renamed and generalized with its new dual functionality.