SearchRecordsPage and ViewModel built upon
Added new bindings for the SearchRecordsPage. Selected record types using the checkboxes now automatically changes and updates the available choices for categories to search for.

Bound the search button to a new command.

Bound the Itemsource for the ListView to a new parameter that is set when the search button is pressed.