SelectRecord page xaml and code behind created
The xaml and code behind of the SelectRecord page is now created. None of the visual elements have been bound to its view model.

its view model hasnt been created yet. but this will be the view when you click on a record that is displayed in the search view.