ViewRecordCharts stuffs moving along
The skeleton xaml for the viewrecordcharts page is done. Its view model is also coming along smoothly.

On the page, two pickers are populated. the first is populated with the relevant years that have records. The second is the relevant months in that year that have records filed. The months itemsource is only updated by the view model on choosing a new year. 

The plan is that once you choose a year and then a month, then a bar chart and pie chart breaking down the expenses categorically will be displayed. Hoping to involve something to compare expense to income for that month/year as well, and perhaps show a historical average for expenses and/or categories. But that is in the future.