ViewRecordCharts page updated
ChartView now has binding to the view model, view model now has property for the bar chart. Bar char property is set every time the user selects a valid month.

CharCalculations now has a function to get entries for a bar chart. It just returns an array of entires that are created. Right now all colors are just random static colors, plan on getting better color choices in the future. Also, this barchart function for entires can possibly (probably should be) made to be generic since all the charts take a list of entires and that wont change.