New and Modified Files
New page files for 'NewCategoryPage' and its view model. Also, new model files for CustomCategory objects and modification to the RecordBook object to hold a list of CustomCategories.

Plan is to store custom categories in another table in the database, or a different database? not sure which makes the most sense tbh.