Custom categories displayed on other pages now
Also, when a record is created with a custom category and is saved in the database, then every time on start, that new category is loaded into the custom category list for the record book. 

e.g., if you create a custom category, and dont create a record using that category, and then exit the app, next time you come back, that category doesnt exist.