New Icon Images, Updates to NewRecordPage and ViewModel
New icons added which are used in the latest updates to the NewRecordPage.

Changes made to the NewRecordPage and its ViewModel. These changes were more so experimental, but a good learning experience. I am not good when it comes to designing UI, so this one is clunky, but it looks better than it used to. These changes very well might replaced in the future for a better UI (i sure hope so)